backend/requirements.txt
# Core
Django==6.0.3
python-decouple==3.8
psycopg2-binary==2.9.11
Pillow==12.1.1

# REST API
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.2

# Background Tasks
celery==5.6.2
redis==7.3.0
django-celery-beat==2.9.0

# AI / LLM
openai==2.26.0
langchain==1.2.12
langchain-core==1.2.18
